Change test from estudios_socioeconomicos to administracion
