--- estudios_socioeconomicos/urls.py
@@ -1,1 +1,0 @@
-
