--- estudios_socioeconomicos/views.py
@@ -1,1 +1,0 @@
-
